# Improve performance of nearby_positions for graphspace