@Tortar 2023-02-24
Given that it is an AbstractGraph on which we are operating probably we need to be for generic in index operation, is this true?